# openapi/nivalis.openapi.yaml
openapi: 3.1.0
info:
  title: Nivalis API
  version: 0.6.0
  description: >-
    Frontend-independent contract for profiles and composable Dashboard state.
    Phase 5 adds API-owned Owner authentication, encrypted Provider credential lifecycle,
    the first real read-only Provider module, and Worker-owned QR/SMS credential acquisition
    for NetEase Cloud Music while preserving
    immutable Dashboard revisions and asynchronous Projection isolation. The API returns
    data and semantic capabilities, never UI styling or Provider secrets.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
security:
  - NivalisSession: []
...
paths:
...
  /v1/public/profile:
    get:
      operationId: getPublicProfile
      summary: Read the public profile
      tags: [Public]
      security: []
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        default:
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/public/dashboards/about:
    get:
      operationId: getPublicAboutDashboard
      summary: Read the published About Me Dashboard snapshot
      tags: [Public]
      security: []
      responses:
        "200":
          description: Published Dashboard read model
          headers:
            ETag:
              $ref: "#/components/headers/ViewETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardReadModel"
        default:
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
  /v1/me/dashboards/about/draft:
    get:
      operationId: getAboutDashboardDraft
      summary: Read the owner's current Draft Dashboard state
      tags: [Dashboard]
      description: Returns immutable owner-controlled configuration only; live Provider data is a separate resource.
      responses:
        "200":
          description: Current Draft Dashboard state
          headers:
            ETag:
              $ref: "#/components/headers/RevisionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardState"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    put:
      operationId: replaceAboutDashboardDraft
      summary: Save the owner's draft without publishing it
      tags: [Dashboard]
      description: Creates an immutable successor revision and atomically advances the Draft pointer.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardDraftUpdate"
      responses:
        "200":
          description: Newly created current Draft revision
          headers:
            ETag:
              $ref: "#/components/headers/RevisionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardState"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/RevisionConflict"
        "422":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/me/dashboards/about/data:
    get:
      operationId: getAboutDashboardLiveData
      summary: Read live projections hydrated for the current Draft configuration
      tags: [Dashboard]
      description: Provider-derived data changes independently from the Draft Revision and its rev ETag.
      responses:
        "200":
          description: Current Draft live Widget data
          headers:
            ETag:
              $ref: "#/components/headers/DataETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardLiveData"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/me/dashboards/about/publish:
    post:
      operationId: publishAboutDashboard
      summary: Explicitly publish the current Draft state
      tags: [Dashboard]
      description: Copies the current Draft to Published inside one database transaction.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Current Published Dashboard state
          headers:
            ETag:
              $ref: "#/components/headers/RevisionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardState"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/RevisionConflict"
        "422":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/me/dashboards/about/widgets:
    post:
      operationId: createAboutDashboardWidget
      summary: Add a Widget instance to the About Me Dashboard draft
      tags: [Widget]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWidgetInput"
      responses:
        "201":
          description: Widget configuration created in a new Draft revision
          headers:
            Location:
              schema:
                type: string
            ETag:
              $ref: "#/components/headers/RevisionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetConfiguration"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/RevisionConflict"
        "422":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        default:
          $ref: "#/components/responses/Problem"
  /v1/me/widgets/{widgetId}:
    parameters:
      - $ref: "#/components/parameters/WidgetId"
    patch:
      operationId: updateWidget
      summary: Update the semantic configuration of a Widget instance
      tags: [Widget]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/UpdateWidgetInput"
      responses:
        "200":
          description: Updated Widget configuration
          headers:
            ETag:
              $ref: "#/components/headers/RevisionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetConfiguration"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/RevisionConflict"
        "422":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteWidget
      summary: Remove a Widget instance from the Dashboard draft
      tags: [Widget]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: Widget instance removed
          headers:
            ETag:
              $ref: "#/components/headers/RevisionETag"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/RevisionConflict"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        default:
          $ref: "#/components/responses/Problem"
  /v1/me/dashboards/about/revisions:
    get:
      operationId: listAboutDashboardRevisions
      summary: List immutable About Dashboard revision metadata
      tags: [Dashboard]
      description: Returns newest-first metadata without loading revision Widget snapshots.
      parameters:
        - $ref: "#/components/parameters/RevisionLimit"
        - $ref: "#/components/parameters/RevisionCursor"
      responses:
        "200":
          description: Cursor-paginated immutable revision history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardRevisionList"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
  /v1/me/providers/netease:
    get:
      operationId: getNeteaseProviderConnection
      summary: Read secret-free NetEase Cloud Music connection state
      tags: [Provider]
      responses:
        "200":
          description: NetEase connection and credential status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderConnection"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    NivalisSession:
      type: apiKey
      in: cookie
      name: nivalis_session
      description: Opaque Nivalis session token; PostgreSQL stores only its hash.
...
  headers:
...
    ViewETag:
      description: Strong validator for Published Revision plus selected live Projection and effective freshness versions.
      schema:
        type: string
        pattern: '^"view:[0-9a-f]{64}"$'
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
...
    Profile:
      type: object
      additionalProperties: false
      required: [displayName, handle, headline, bio, avatarUrl, tags]
      properties:
        displayName:
          type: string
        handle:
          type: string
        headline:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        tags:
          type: array
          items:
            type: string
...